steps in right direction for registry but still not working
